Refactor : 검색 컴포넌트와 헤더 컴포넌트 분리 및 코드 수정
